Move TailwindEntity into new file
It looks like some work was started to create a parent, but this is continuing that.
